Add files via upload
The expanded use of the automation system has led to its use in many fields and is now also 
used in the education infrastructure. It is used by many universities, but no one has yet a 
complete solution to every problem and task. The goal of this project is to develop a web 
application for university automation. This allows you to do all the work of the university, even 
if your staff doesn't have a lot of throughputs. Every university today consists of a database 
that stores data about each student, Cloud computing technologies are thus very useful in 
storing information. Cloud computing is an extension of a paradigm where the competencies 
of applications are exposed as services. These services empower the development of scalable 
web applications in which dynamically scalable and often virtualized resources are provided 
as a service over the Internet. University automation system consists of modules like student, 
faculty and admin.